backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2

# Document processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# AI
anthropic==0.18.1

# Web scraping
beautifulsoup4==4.12.3
aiohttp==3.9.3
lxml==5.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
cryptography==42.0.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
